apolo_visual_rag/requirements.txt
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
transformers==4.46.1
pdf2image==1.17.0
lancedb==0.15.0
ipython==8.29.0
typer==0.12.5
colpali_engine==0.3.3
streamlit==1.39.0
openai==1.53.0
pypdf==5.1.0
ruff==0.7.1
polars==1.12.0
streamlit-pdf-viewer==0.0.19